Move deprecation code from xpack core to deprecation module. (#74120)
This reduces the number of imports and all deprecation code is now in one place.
Also removed DeprecationChecksTests test case from x-pack core, because the same test already exists in the deprecation module.
